登入/註冊頁面樣式調整，使用sweetalert

+++ scripts/login.js
@@ -9,6 +9,53 @@ function logIn(a, b) {
     'password': b,
   }).then(function (response) {
     console.log(response)
+
+    localStorage.setItem('token', response.data.accessToken);
+    localStorage.setItem('usersId', response.data.user.id);
+    
+    logInForm.reset();
+
+    // 登入成功提示
+    Swal.fire({
+      icon: "success",
+      title: "登入成功",
+      showConfirmButton: false,
+      timer: 1500
+    }).then(res => {
+      window.location.href = '/findSpace.html';
+    })
+
+  }).catch(function (err) {
+    console.log(err);
+    hideError('.email-warn');
+    hideError('.Password-warn');
+
+    // 錯誤信箱格式提示
+    if (err.response.data === 'Email format is invalid') {
+      showError('.email-warn')
+    }
+    // 錯誤密碼提示
+    else if (err.response.data === 'Incorrect password') {
+      showError('.Password-warn')
+    }
+    // 無帳號提示
+    else if (err.response.data === 'Cannot find user'){
+      Swal.fire({
+        icon: "error",
+        title: "登入失敗",
+        text: "請確認您已註冊"
+      });
+    }
+  })
+}
+
+// 芒果樹原版
+/* function logIn(a, b) {
+  axios.post(Url + '/login', {
+    'email': a,
+    'password': b,
+  }).then(function (response) {
+    console.log(response)
     const token = response.data.accessToken;
     const usersId = response.data.user.id;
     localStorage.setItem('token', token);
@@ -29,28 +76,35 @@ function logIn(a, b) {
     }
     if (err.response.data === 'Incorrect password') {
       showError('.Password-warn')
+      return
     } else if (email !== '') {
       hideError('.Password-warn')
     }
-    alert('登入失敗');
+    else{
+      Swal.fire({
+        icon: "error",
+        title: "登入失敗",
+        text: "請確認已註冊"
+      });
+    }
   })
-}
+} */
 
 logInBtn.addEventListener('click', function (e) {
   let email = emailInput.value.trim();
   let passWord = passWordInput.value.trim();
+
+  hideError('.email-warn');
+  hideError('.Password-warn');
+
   if (email === '') {
-    showError('.email-warn')
-  } else if (email !== '') {
-    hideError('.email-warn')
+    showError('.email-warn');
   }
   if (passWord === '') {
-    showError('.Password-warn')
-  } else if (passWord !== '') {
-    hideError('.Password-warn')
+    showError('.Password-warn');
   }
   if (email !== '' && passWord !== '') {
-    logIn(email, passWord)
+    logIn(email, passWord);
   }
 })
